Adding user profile data and groups from custom schema.
Warning: This is work in progress and still needs clean up.

Additional information is retrieved from the user profile and can be passed to NGINX variables.
- full name
- username (first part of email, like before)
- email

And groups are fetched from Google Directory User customSchemas.
A service account with domain-wide-delegation was setup for the access and it needs to impersonate a Google Suite Admin.